requirements.txt
# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Market data
yfinance>=0.2.0
ccxt>=4.0.0

# Database (Supabase)
supabase>=2.0.0

# Security
PyJWT>=2.8.0
python-dotenv>=1.0.0

# Data visualization
plotly>=5.18.0

# Machine Learning
scikit-learn>=1.3.0
torch>=2.0.0

# HTTP
requests>=2.31.0
httpx>=0.25.0

# Schema validation
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
